Extract action log entry using struct (WIP)
