+++ pkg/testfunctional/computednestedlist/test_resource_plugin_framework_computed_nested_list.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/testfunctional/actionlog"
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
@@ -26,34 +27,6 @@ type computedNestedListResourceModelV0 struct {
 	Id         types.String `tfsdk:"id"`
 }
 
-type ActionLogEntry struct {
-	Action types.String `tfsdk:"action"`
-	Field  types.String `tfsdk:"field"`
-	Value  types.String `tfsdk:"value"`
-}
-
-func getActionLogEntrySchema() map[string]schema.Attribute {
-	return map[string]schema.Attribute{
-		"action": schema.StringAttribute{
-			Required: true,
-		},
-		"field": schema.StringAttribute{
-			Required: true,
-		},
-		"value": schema.StringAttribute{
-			Required: true,
-		},
-	}
-}
-
-func getActionLogEntryTypes() map[string]attr.Type {
-	return map[string]attr.Type{
-		"action": types.StringType,
-		"field":  types.StringType,
-		"value":  types.StringType,
-	}
-}
-
 func (r *computedNestedListResource) Metadata(_ context.Context, request resource.MetadataRequest, response *resource.MetadataResponse) {
 	response.TypeName = request.ProviderTypeName + "_computed_nested_list"
 }
@@ -72,7 +45,7 @@ func (r *computedNestedListResource) Schema(_ context.Context, _ resource.Schema
 			},
 			"actions_log": schema.ListNestedAttribute{
 				NestedObject: schema.NestedAttributeObject{
-					Attributes: getActionLogEntrySchema(),
+					Attributes: actionlog.GetActionLogEntrySchema(),
 				},
 				Computed: true,
 			},
@@ -112,7 +85,7 @@ func (r *computedNestedListResource) Create(ctx context.Context, request resourc
 }
 
 func setActionsOutputThroughStruct(ctx context.Context, response *resource.CreateResponse, data *computedNestedListResourceModelV0) {
-	var actions []ActionLogEntry
+	var actions []actionlog.ActionLogEntry
 	diags := data.ActionsLog.ElementsAs(ctx, &actions, false)
 	if diags.HasError() {
 		response.Diagnostics.Append(diags...)
@@ -130,12 +103,12 @@ func setActionsOutputThroughStruct(ctx context.Context, response *resource.Creat
 func setActionsOutputExplicit(ctx context.Context, response *resource.CreateResponse, data *computedNestedListResourceModelV0) {
 	existingEntries := data.ActionsLog.Elements()
 
-	actions := make([]ActionLogEntry, 0)
-	actions = append(actions, actionEntry("SOME ACTION", "ON FIELD", "WITH VALUE"))
-	actions = append(actions, actionEntry("SOME OTHER ACTION", "ON OTHER FIELD", "WITH OTHER VALUE"))
+	actions := make([]actionlog.ActionLogEntry, 0)
+	actions = append(actions, actionlog.ActionEntry("SOME ACTION", "ON FIELD", "WITH VALUE"))
+	actions = append(actions, actionlog.ActionEntry("SOME OTHER ACTION", "ON OTHER FIELD", "WITH OTHER VALUE"))
 
 	for _, a := range actions {
-		entry, diags := types.ObjectValue(getActionLogEntryTypes(), map[string]attr.Value{
+		entry, diags := types.ObjectValue(actionlog.GetActionLogEntryTypes(), map[string]attr.Value{
 			"action": a.Action,
 			"field":  a.Field,
 			"value":  a.Value,
@@ -147,18 +120,10 @@ func setActionsOutputExplicit(ctx context.Context, response *resource.CreateResp
 		existingEntries = append(existingEntries, entry)
 	}
 	var diags diag.Diagnostics
-	data.ActionsLog, diags = types.ListValueFrom(ctx, types.ObjectType{AttrTypes: getActionLogEntryTypes()}, actions)
+	data.ActionsLog, diags = types.ListValueFrom(ctx, types.ObjectType{AttrTypes: actionlog.GetActionLogEntryTypes()}, actions)
 	if diags.HasError() {
 		response.Diagnostics.Append(diags...)
 		return
-	}
-}
-
-func actionEntry(action string, field string, value string) ActionLogEntry {
-	return ActionLogEntry{
-		Action: types.StringValue(action),
-		Field:  types.StringValue(field),
-		Value:  types.StringValue(value),
 	}
 }
